add a condition on minibuffer in (bfs-environment-is-corrupted-p) function
Sometimes using the minibuffer triggers the functions
in (window-configuration-change-hook).  We don't want to exit the
minibuffer or have message showing up in the minibuffer when using it
when "bfs" environment is turned on.  So we
tell (bfs-environment-is-corrupted-p) to return nil value when
emacs run (window-configuration-change-hook) hooks when we are
doing stuff in the minibuffer.

* bfs.el:  Add a condition on minibuffer
  in (bfs-environment-is-corrupted-p) function.