update (bfs-backward-update-visited) command
The cursor in child buffer can be on a non readable file,
and so leads the function (bfs-child-entry) to return the file name
of a non readable file.  As non readable files can't be visited
with (bfs-preview), we don't want to store pairs (PARENT . CHILD-ENTRY)
where CHILD-ENTRY isn't readable in (bfs-backward-last-visited) variable.

* bfs.el:  Update (bfs-backward-update-visited) command.