simplify signature of (bfs-update, bfs-update-backward-visited) functions
In the future (bfs-update) might be a function user would want to use
to build their own commands that select a file with any completion
style they want and update the frontend accordingly with the file
selected.  So we simplified the (bfs-update) API.
Before, (bfs-update) took 2 arguments (parent child-entry), now it
takes only one argument (child) that is simply a file path.

(bfs-backward-visited) list was build as an alist of (PARENT . CHILD-ENTRY).
Now it is build as a list of file paths.

* bfs.el:  Update (bfs-update, bfs-forward, bfs-find-file,
  bfs-update-backward-visited, bfs-get-backward-visited).