; fix (bfs-next) command
We changed how we insert the list of directory in child buffer.  We
insert it with (bfs-insert-ls) and we do not insert "\n" at the end
of last word of child buffer.  So the number of line in child buffer
(also in parent buffer) is reduced by one.  So this fix.

* bfs.el:  Fix (bfs-next) command.