no longer use (font-lock-mode) to add text properties (face) in (bfs-mode)
(font-lock-mode) is interesting when you don't know when the user will
add text that need to be colored (typically in prog-mode).  But
in (bfs-mode) when we insert text in (child) and (parent) we already
have the necessary information to add text property to the text.  So
now we do.  As consequence, we don't need to add a trailing slash "/"
to directories in order to differenciate them from normal file and so
add them text property face with font-lock-mode.

* bfs.el:  No longer use (font-lock-mode) to add text
  properties (face) in (bfs-mode). Update
  (bfs-ls-group-directory-first) function.  Remove (bfs-re-dir,
  bfs-font-lock-keywords) variable.  Add (bfs-file, bfs-file-face)
  variables.  Update (bfs-mode) function.