add (bfs-preview-buffer-name) variable, add (bfs-preview-buffer) function
There are some files that we don't want to preview.  We will
use (bfs-preview-buffer-name) buffer to give us visual feedback in the
normal preview window when we move in the child window (for instance
with (bfs-next) or (bfs-previous)).  It is more intuitive for me
than nothing happening or just in the echo area, because when I use
bfs movement command I always expect something to happened in the
preview window.

* bfs.el:  Add variable (bfs-preview-buffer-name).  Add local
  variable (bfs-preview-buffer-file-name).  Add (bfs-preview-buffer)
  function.  Update (bfs-clean).