rename and modify (bfs-file-readable-p => bfs-valid-child-p) function
As (bfs-preview) can use a preview buffer to signaling when a file
can't be viewed, we don't have to limit so much which file path can
be a valide child.  A file path can be a bfs child if it exist and
isn't root path.  This is more flexible and I think more user
friendly.

* bfs.el:  Remove (bfs-child-is-valid-p) function
  Rename and modify (bfs-file-readable-p => bfs-valid-child-p) function.