simplify the code by changing values returned by (bfs-child) function
Now (bfs-child) return nil when the cursor in child buffer is on an
empty line.  This occurs when after filtering, the child buffer is
left empty, or when the cursor is at end of buffer (the last line).
Doing this, the (bfs-child) function gives us another information on
the state of bfs environment that we can benefit from and use.
It seems to simplify some part of the code.

* bfs.el:  Modify the value return by (bfs-child).  Accordingly update
  the functions:
  bfs-update-visited-backward,
  bfs-backward,
  bfs-forward,
  bfs-top-mode-line,
  bfs-preview-matches-child-p,
  bfs-top-buffer,
  bfs-preview-buffer,
  bfs-preview,
  bfs-update,
  bfs-check-environment.
  Accordingly update the variable (bfs-do-not-check-after).

bfs-do-not-check-after