remove leading space of bfs specific buffers
What it is great with leading space in buffer name, is that they are
not present in the user (buffer-list).  But it seems
that (font-lock-mode) doesn't work with buffers that have a name
starting with a leading space.  As we want the user, to be able to use
font-lock-mode to add any keywords needed, we choose remove the
leading space of bfs specific buffers.  If we think it is a good idea,
maybe in the future we will burry those bfs specific buffers.

* bfs.el:  Update (bfs-top-buffer-name, bfs-parent-buffer-name,
  bfs-child-buffer-name, bfs-preview-buffer-name).