add font lock keywords variables to bfs modes
If you don't define a variable holding the keywords
for (font-lock-mode) and don't make the keywords value
of (font-lock-defaults) variable in the mode you define,
users can't add there own keywords for (font-lock-mode) outside the
hook of the mode.  So we define those keywords variables (set to nil),
for the (bfs-top-mode, bfs-preview-mode, bfs-parent-mode, bfs-mode).

* bfs.el:  Add variables (bfs-top-font-lock-keywords,
  bfs-preview-font-lock-keywords, bfs-parent-font-lock-keywords,
  bfs-font-lock-keywords).  Set (font-lock-defaults) variables in
  (bfs-top-mode, bfs-preview-mode, bfs-parent-mode, bfs-mode).