remove (bfs-file) face
By not using (bfs-file) face to fontify entries in child and parent
buffers that represent files, those entries have no faces, so use
the (default) face.  This way, users can use (font-lock-mode) to
fontify those entries without using the override flag
of (font-lock-keywords).  This gives us the possibility (in the future)
to mark entries and changes the face accordingly (this is not the
case if a font lock keyword has a face with the override flag).

* bfs.el:  Remove (bfs-file) face.  Update accordingly
  (bfs-top-child-entry, bfs-top-symlink-file-target) faces to inherit
  the (default) face.