; fix line highlighting in child buffer
`bfs-line-highlight-child` function was used before the buffer
was totally fontify by font-lock, so it was using the wrong faces
to highlight line in the child buffer.  We tried to delay the moment
where `bfs-line-highlight-child` was call with `run-at-time` function,
but it wasn't reliable.

Now we have a better solution.  We directly fontify the buffer by
calling `(font-lock-fontify-buffer)` and the we call
`bfs-line-highlight-child`.

* bfs.el: remove `bfs-line-highlight-child-delay` variable.  Update
  `bfs-child-buffer`.