refactor (bfs-mode, bfs-insert-ls-child, bfs-child-buffer)
we've made this refactoring to set `bfs-max-length` at the same
place it is used.  So if how we calculate this value change, now it
change only in one place.

As turn on (bfs-mode) removes all the local variable, we add to change
the place where we turn it on, to be able to set correctly
`bfs-max-length` without overriding it.

Due to those changes, we've change the place where we highlight the
current entry line.  At this place, as the buffer is already seen as
fontify but we want to sure it is fontify correctly, we no longer the
function `(font-lock-fontify-buffer)` but the function
`font-lock-ensure`.

* bfs.el: refactor (bfs-mode, bfs-insert-ls-child, bfs-child-buffer).