refactor (bfs-parent-mode, bfs-insert-ls-parent, bfs-parent-buffer)
we've made this refactoring to set `bfs-max-length` at the same
place it is used.  So if how we calculate this value change, now it
change only in one place.

As turn on (bfs-parent-mode) removes all the local variable, we add to
change the place where we turn it on, to be able to set correctly
`bfs-max-length` without overriding it.

* bfs.el: refactor (bfs-parent-mode, bfs-insert-ls-parent, bfs-parent-buffer).