update (bfs-find-file, bfs-project-find-file-in, bfs-project-find-file)
As the commands `bfs-find-file` and `bfs-project-find-file`
upadte `bfs` environment they must udpate `bfs-visited-last` list.
This is what we do here.
When we update `bfs` environment choosing a directory, we want the
selected child in this directory to be the last visited if there is
one last visited child in this directory.

* bfs.el:  update (bfs-find-file, bfs-project-find-file-in,
  bfs-project-find-file).