handle errors in (bfs) command to be sure to leave user's emacs state clean
handle errors in (bfs) command to be sure to leave user's emacs state
clean when something goes wrong when using (bfs) command.
When an errors occurs with (bfs) command, bfs environment is cleaned
up and the error is printed in the echo area.  If the
variable (debug-on-error) is non-nil, we enters in the debugger if an
error is signaled.

* bfs.el: update (bfs) command.