add support for (which-key-mode) external packages
The mode `which-key-mode` displays the key bindings following
your currently entered incomplete command (a prefix) in a popup.
As (bfs) takes control over the layout of the frame it uses and
check after each command the validity of the layout, we don't want
`which-key-mode` to pop up anywhere and forces (bfs) environment to be
quitted.  So when (bfs) environment is active we force
`which-key-mode` to use the minibuffer to pop up its information by
temporary setting the variable `which-key-popup-type` to 'minibuffer.
This way (bfs) and `which-key-mode` can both run together.

* bfs.el:  update (bfs) command.  update (bfs-clean) function.