update (bfs-visited-last-in-dir, bfs-visited-last-push) for better performance
I run the emacs profiler and it seems that the use of `f-equal-p`
function in the context of the function (bfs-visited-last-in-dir,
bfs-visited-last-push) takes CPU usage elevated in comparison of
other functions and command used in bfs environment (except
`find-file-noselect`).

As we have control over the production of `bfs-visited-last` variable,
we know that we can compare the strings and not file that are
represented by the string (in that case).  So we prefer using
`string=` function that `f-equal-p`.

Maybe I'm wrong.  We'll see.

Another (feeling), it seems that locally defining the function
`dirname=` in order to use it in `--remove` macro is "faster".

* bfs.el:  update (bfs-visited-last-in-dir, bfs-visited-last-push).